pyproject.toml
[tool.poetry]
name = "santander-bootcamp-ciencia-de-dados"
version = "0.1.0"
description = ""
authors = ["eduardohealli <eng.eduardohealli@gmail.com>"]
packages = [{include = "config"}]

[tool.poetry.dependencies]
python = "^3.10"
seaborn = "^0.12.2"
pandas = "^2.1.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
